+++ pages/authCallbackShib/authCallbackShib.js
@@ -7,6 +7,7 @@ var csrf = require('../../lib/csrf');
 var sqldb = require('@prairielearn/prairielib/sql-db');
 
 router.get('/:action?/:target(*)?', function(req, res, next) {
+    if (!config.hasShib) return next(new Error('Illinois Shibboleth login is not enabled'));
     var authUid = null;
     var authName = null;
     var authUin = null;