+++ pages/authLogin/authLogin.js
@@ -10,14 +10,11 @@ var config = require('../../lib/config');
 var sql = sqlLoader.loadSqlEquiv(__filename);
 
 router.get('/', function(req, res, next) {
-    res.locals.showAll = true;
-
     var params = {
         ip: req.ip,
         force_mode: (config.authType == 'none' && req.cookies.pl_requested_mode) ? req.cookies.pl_requested_mode : null,
         req_date: res.locals.req_date,
     };
-
     sqldb.query(sql.get_mode, params, function(err, result) {
         if (ERR(err, next)) return;
         res.locals.mode = result.rows[0].mode;