+++ ChangeLog.md
@@ -45,6 +45,8 @@
 
   * Fix "Logout" button with Google authentication (Matt West).
 
+  * Fix error message when an authentication method is not enabled (Matt West).
+
   * Change `pl-code` to display code from a source file (Mariana Silva).
 
   * Change element names to use dashes instead of underscores (Nathan Walters).